refactor: 유저 스토어에 저장
