+++ src/libs/store/modules/index.js
@@ -1,5 +1,4 @@
 import { combineReducers } from 'redux';
 import user from './user';
-import signup from './signup';
 
-export default combineReducers({ user, signup });
+export default combineReducers({ user });
